detect/asn1: require packet during setup
Following d3ccff5822, in master, an ASN.1 rule is evaluated as a pkt
type rule. In 7.0.x, the rule defaults to a stream type rule as it is
added to the payload SigMatch list.
A rule's evaluation can change based on its type so this inconsistency
may cause different behavior for asn1 rules in 7.0.x and master. Fix this
by making signature require packet.

Bug TODO