smtp: fix the point of triggering stream inspection
As a part of ab01a1b2, in order to match the behavior in master, the
calls for triggering raw stream inspection were made when communication
in one direction for a transaction was completed. However, it was
incorrect to do so. Reliable inspection requires any request/response to
be completed. Hence, fix the behavior and make it consistent with all
the other applayer parsers.

SMTP parser can handle multiple command lines per direction. Appropriate calls
to trigger raw stream inspection have been added on succesful parsing of each
request and response line.

Bug 7783