flowbits: add a validation callback during setup
This should make it possible to catch invalid combinations in the same
signature early. This patch covers checking the following invalid cmd
combinations:
- set + isset
- unset + isnotset
- set + toggle
- set + unset
- isset + isnotset

the same flowbit in the same signature which is basically an unnecessary
operation.
This helps bring down the difficulty of handling of actual complex flowbit chains.

Bug 7772
Bug 7773
Bug 7774
Bug 7817
Bug 7818