detect/entropy: Unique flowvar names
Issue: 7814

Support unique flow variable names for entropy-related values.

Since this is a breaking change from Suricata 8, unique names
are configurable (default value on) with this setting

        logging:
          entropy:
            make-unique: on

When `make-unique` is on, entropy-related flow variable names are
structured as:
     <sticky_buffer>_<sid>_<occurrence>

Occurrence is a small integer value corresponding to the number of
entropy values in the rule. The first entropy usage will create a flow
variable with an occurrence value of 1; the second will use 2, and so
on. This is necessary to disambiguate the flow variable names.

When `make-unique` is off (default = on), this format is used:
     <sticky_buffer>