file: mime type as alternative to magic
File magic is known to have performance issue. When looking for
an alternative, I've been pointed to the tree_magic_mini crate
that output the mime type of a binary stream. This is different
from magic but it has multiple advantages as it is a standard
so it can be correlated with other tools.

So instead of replacing magic, this patch adds a new mime type
output to fileinfo events.

This patch also adds the `file.mimetype` sticky buffer.

Mime type has 2 advantages over file magic. First it is really
faster and second, the result are easier to use as the MIME type
are well defined. But it provides less information than magic for
example with regards to the size of images.

This patch is using version 3.2.0 of tree_magic_mini that provides
capabilities to load mime files from a give directory.

This patch is embedding the GPL database if an explicit flag is
passed to configure (--enable-bundled-gpl-mimetype). In this case
the mime type files comes from `tree_magic_mini` dependencies.

The patch adds the mime definition in Suricata source code.
This means that if the user uses this data, then we have no dependency
in term of results on the operating system and on the evolution of
the crate. Updating the mime type definition on the target system
will be done by installing the files present in Suricata code.

Installation of the mime type files is done via `make install-conf`
to the mimetype/ directory below data directory. This directory is
used by default even if the suricata.yaml is not updated with the
new `mimetype-dir` config variable.

The new Makefile target `upgrade-data` takes care of upgrading
the file. Users will be able to use it when there is a new version
available.

The `mimetype-dir` YAML variable can be used to set up a different
directory with data under control by the user.

Ticket: 7816