eve: log app_proto in all event types
There was a regression between Suricata 7 and Suricata 8. The
app_proto was logged in almost all events in 7 and is only log
in a small subset (fileinfo, flow, frame, netflow) in 8.

This patch updates the code to log app_proto in all events if
there is a Flow available. It is making use of EveAddAppProto
function to get interesting information such as original
application protocol or difference between server and client
side.

Ticket: #7888