eve: log app_proto in all event types
The app_proto is logged in a small subset (fileinfo, flow,
frame, netflow) in Suricata when it is an interesting information
for all events type as it includes detection information and protocol
transition.

This patch updates the code to log app_proto in all events if
there is a Flow available. It is making use of EveAddAppProto
function to get interesting information such as original
application protocol or difference between server and client
side.

Backward compatibility is preserved as app_proto information
will only be logged when the eve-log.version is greater or
equal to 2.

Ticket: 7888