ike: suffix duplicate attributes with an index
The attributes in a proposal are an array so its possible to have
duplicate values, even if not within spec.

To address this in a backwards compatible way, suffix duplicates with
their index. The first seen attribute will not have a suffix, but the
first duplicate attribute type will have suffix 1, etc.

For example:

  "client": {
    "proposals": [
      {
        "sa_life_duration": "Unknown",
        "sa_life_duration_raw": 86400,
        "sa_life_duration_1": "Unknown",
        "sa_life_duration_1_raw": 65535
      }
    }
  }

Ticket: #7923