app-layer: add partial ok return code
Allow a parser to return OK with only consuming partial data and then
immediately processing the rest of the data, giving it a change to rerun
protocol detection on the remaining data.

For this a new AppLayerResult::ok_partial_continue has been added, which
only indicates the number of bytes consumed.

Meant for SSLproxy, where the data will include the special header as well
as the first line of the decrypted protocol. The SSLproxy parser will then
be able to return after the special header and have the data for it consumed.