flow: decrypted flow support
Allow tagging of a flow as decrypted. This also allows the caller to set
a new flow tuple that will be used to translate new packets to the new
tuple.

This is developed for the SSLproxy usecase, where the packets on the
wire do not reflect the real tuple, but instead a connection between the
proxy and a helper tool. The SSLproxy header at the start of the
connection holds the original tuple, so that can then be applied to the
packets.