main/options: Limit verbose option count
Issue: 7389

Continue to recognize multiple 'v' specifications up to the maximum
recognized. The maximum recognized is the number of slots between the
NOTICE and DEBUG levels.

When debug is configured, `-vvvv` is required to set the log level to
debug. Specifying 4 or more `v` will maintain debug level.

When debug is not configured, `-vvvv` and more `v`'s will cause
config messages to be printed.