affinity: skip affinity unit tests on incompatible environments
Some unit tests work with multiple CPUs. Suricata, in some code paths,
is dependant on the environment on which it runs. As a result, e.g.
tests with 4 cores failed on 2 CPU core machines.
This commit introduces new skip directive for these tests to be not
evaluated.

Part of this work required removing the check in the parsing logic
as it failed the tests even before it got to the skip directive.
(The b > max check in src/util-affinity.c)

Now, the validation check moves to after the full affinity config
is loaded. This brings more in-depth check as previously it only checked
upper bound of the range definition ("a-B") in the cpu set.

AffinityCPUConfigIsCompatible now checks for both lower and upper bounds
of the range check, as well as, the list of individual CPUs. In other
words, this post-load check is not dependent on the particular CPU set
definition.

Ticket: 6927