snmp: can be set to detection-only
Realloc alp_ctx.ctxs when a dynamic alproto is registered and
g_alproto_max increases. So dynamic alproto can be treated as
real/normal ones. And app-layer switch can be set to any value
of no/deteciton-only/yes.
Call AppLayerProtoDetectRegisterProtocol in func
SCPluginRegisterAppLayer, Because detection-only is implicitly
enabled for an app-layer plugin.

Ticket: 8000