transform/subslice: Add subslice transform
The subslice transform creates a slice of the input buffer.

Specify the subslice desired -- nbytes is optional:
        subslice: [offset <,nbytes>]

offset: Specifies the starting offset for the new subslice. When
negative, expresses how far from the end of the input buffer to begin.

nbytes: Specifies the size of the subslice. When negative, specifies the
byte count preceding the offset to include.

When nbytes is not specified, the size of the subslice will be the size
of the input buffer - offset.

Examples:
        subslice[1]     - The subslice will be a copy of the input
            buffer but omit the input buffer's first byte
            "This is Suricata" -> "his is Suricata"
        subslice[0, 13] - The slice is created from the first 13 bytes
            of the input buffer
            "This is Suricata" -> "This is Suric"
        subslice[10, -5] - The subslice is created starting at offset 10
            and continuing to 5 bytes before the end. This is the same as
            subslice[5, 5]
            "This is Suricata" -> "is Su"
        subslice[-3] - The subslice will be the last 3 bytes of the
            input buffer.
            "This is Suricata" -> "ata"